\name{print.SurvMTP}
\alias{print.SurvMTP}
\title{Print an object of class SurvMTP--output from survMTP.estimate}

\description{
  S3 method for class ''SurvMTP''
}
\usage{
#S3 method for class ''trtsel''
\method{print}{SurvMTP}(x, ...)
}

\arguments{
  \item{x}{output from the function survMTP.estimate}
  \item{\dots}{NA}
}
\value{
  NULL
}

