\name{print.SurvMTP_cch}
\alias{print.SurvMTP_cch}
\title{Print an object of class SurvMTP_cch--output from survMTP.cch}

\description{
  S3 method for class ''SurvMTP_cch''
}
\usage{
#S3 method for class ''trtsel''
\method{print}{SurvMTP_cch}(x, ...)
}

\arguments{
  \item{x}{output from the function survMTP.cch}
  \item{\dots}{NA}
}
\value{
  NULL
}

