\name{print.SurvMTP_ncc}
\alias{print.SurvMTP_ncc}
\title{Print an object of class SurvMTP_ncc--output from survMTP.ncc}

\description{
  S3 method for class ''SurvMTP_ncc''
}
\usage{
#S3 method for class ''trtsel''
\method{print}{SurvMTP_ncc}(x, ...)
}

\arguments{
  \item{x}{output from the function survMTP.ncc}
  \item{\dots}{NA}
}
\value{
  NULL
}

